+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NeedToKnow
 
+### v4.7.7 (2024-08-22)
+* Fixed: Buffs and debuffs tracked by spell ID on target/focus should now work again
+* Changed: Default profile should now provide cleaner experience for first-time users and new characters
+
 ### v4.7.6 (2024-08-21)
 * Fixed: Bars for equipment slot cooldowns can now show icon
 * Fixed: Tracking buffs/debuffs on target of target should work again